WIP Working through fix for 7807
It appears that there is something where cbs_h264_read_nal_unit() is being called too early, or is somehow otherwise being called with the wrong segment of bitstream. The reason this is probably happening may have to do with the redundant pps mentioned in the ticket, but I'm not yet sure where that is happening exactly.