WIP Temporarily turn off some logging for golomb decoder, turn on some other logging in ffmpeg
It appears that the 'failure' of the Exp-Golomb decoder is not necessarily directly connected to the failure to output the video. That seems more connected to the h264_redundant_pps_filter which is the type of filter that's failing at the moment, in the provided example execution. This can be seen in that with new logging added we are able to see the name of that filter, and it is h264_redundant_pps_filter.